backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
beautifulsoup4==4.12.3
lxml==5.3.0
urllib3==2.6.3
playwright==1.55.0
pdfplumber==0.11.4
openpyxl==3.1.5
openai==1.108.0
pgvector==0.3.6
tiktoken==0.8.0
